backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
sse-starlette==2.1.0
python-multipart==0.0.12

# Async support
httpx==0.27.2
aiohttp==3.10.10

# Pydantic for schemas
pydantic==2.9.2
pydantic-settings==2.5.2

# Azure SDKs
azure-identity==1.19.0
azure-storage-blob==12.23.1
azure-cosmos==4.7.0
azure-search-documents==11.6.0b7
openai==1.54.3

# Microsoft Agent Framework
agent-framework>=0.1.0

# Redis for event streaming
redis==5.2.0

# Database
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.36

# Observability
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-exporter-otlp==1.28.1
azure-monitor-opentelemetry==1.6.3

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
